Moves sync implementation into extra file
